backend/python/python-ai/web-sercher/requirements.txt
flask
flask-session
python-dotenv
crewai
requests
bootstrap-flask